% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting.R
\name{add_splits}
\alias{add_splits}
\title{construct rsample splits from data in an environment}
\usage{
add_splits(env, ...)
}
\arguments{
\item{env}{environment}

\item{...}{arguments passed to initial_split}
}
\description{
construct rsample splits from data in an environment
}
