% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{analysis}
\alias{analysis}
\title{Object specification of the ML pipeline}
\format{
A list containing three elements
\describe{
\item{splits}{an \code{rsplit} object of the \code{rsample} package that indicates splitting criteria and data inclusion}
\item{recipes}{assortment of \code{recipes} package recipe objects describing preprocessing steps}
\item{models}{a list of \code{parsnip} model specifications to see hyperparameter tuning criteria}
}
}
\source{
\url{https://github.com/D-Se/ML}
}
\usage{
analysis
}
\description{
A collection of objects along the critical path of analysis process
useful for model prototyping and verification.
}
\keyword{datasets}
