% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_bookdown.R
\name{change_ext}
\alias{change_ext}
\title{Change extension from md to Rmd}
\usage{
change_ext(file, inext, outext)
}
\arguments{
\item{file}{a Rmarkdown file to convert to regular markdown}

\item{inext}{file extension to replace}

\item{outext}{file extention to replace with}
}
\description{
Change extension from md to Rmd
}
