% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-table.R
\name{coef_table}
\alias{coef_table}
\title{Coefficient Table}
\usage{
coef_table(x)
}
\arguments{
\item{x}{An ml_analysis Object}
}
\value{
A tibble of the coefficient table.
}
\description{
Coefficient Table
}
