% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_plot}
\alias{create_plot}
\title{Plot ozone from the preprocessed air quality data.}
\usage{
create_plot(data)
}
\arguments{
\item{data}{Data frame, preprocessed air quality dataset.}
}
\value{
A ggplot histogram showing ozone content.
}
\description{
Plot a histogram of ozone concentration.
}
