% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidate.R
\name{crossvalidate}
\alias{crossvalidate}
\title{Generalized Cross Validation}
\usage{
crossvalidate(xv_data, h, ...)
}
\arguments{
\item{y}{response variable}

\item{x}{predictors}

\item{r}{number of epochs}

\item{mu}{margin for the margin perceptron}

\item{E}{number of epochs}

\item{dynamic}{boolean value to turn on dynamic learning rate}

\item{averaged}{boolean value to turn on averaging}

\item{has_bias}{if the bias term has been included or not (defaults to FALSE)}

\item{seed}{some value for setting the seed for random number generation}
}
\value{
tibble of the input file
}
\description{
Generalized Cross Validation
}
\examples{
crossvalidate()
}
