% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{data}
\alias{data}
\title{temporary airquality data for bookdown}
\format{
see \code{airquality}
}
\source{
New York State Department of Conservation (ozone data) and the National Weather Service (meteorological data).
}
\usage{
data
}
\description{
temporary airquality data for bookdown
}
\keyword{datasets}
