% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras.R
\name{define_model}
\alias{define_model}
\title{Define a Keras model.}
\usage{
define_model(recipe, units1, units2, act1, act2, act3)
}
\arguments{
\item{recipe}{A \code{recipe} object from \code{\link[=prepare_recipe]{prepare_recipe()}}.}

\item{units1}{Number of neurons in the first layer.}

\item{units2}{Number of neurons in the second layer.}

\item{act1}{Activation function for layer 1.}

\item{act2}{Activation function for layer 2.}

\item{act3}{Activation function for layer 3.}
}
\value{
A Keras model. Not compiled or run yet.
}
\description{
Define a Keras model for the customer churn data.
}
