% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{.build_provenance}
\alias{.build_provenance}
\title{Build provenance metadata from training data for storage in a Model.}
\usage{
.build_provenance(data)
}
\description{
Build provenance metadata from training data for storage in a Model.
}
\keyword{internal}
