% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{.check_provenance}
\alias{.check_provenance}
\title{Check cross-verb provenance. Errors on split-shopping.}
\usage{
.check_provenance(model_provenance, test)
}
\description{
Check cross-verb provenance. Errors on split-shopping.
}
\keyword{internal}
