% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{.decode}
\alias{.decode}
\title{Decode integer predictions back to original labels}
\usage{
.decode(predictions, norm)
}
\description{
Decode integer predictions back to original labels
}
\keyword{internal}
