% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{.encode_target}
\alias{.encode_target}
\title{Encode target vector using stored label_map}
\usage{
.encode_target(y, norm)
}
\description{
Encode target vector using stored label_map
}
\keyword{internal}
