% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{.is_assessed}
\alias{.is_assessed}
\title{Check if a test partition has already been assessed by any model}
\usage{
.is_assessed(df)
}
\description{
Check if a test partition has already been assessed by any model
}
\keyword{internal}
