% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{.transform_fit}
\alias{.transform_fit}
\title{Fit and apply encoding to training features}
\usage{
.transform_fit(X, norm)
}
\description{
Fit and apply encoding to training features
}
\keyword{internal}
