% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance.ml_analysis}
\alias{glance.ml_analysis}
\title{Glance at an object}
\usage{
\method{glance}{ml_analysis}(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{other arguments passed to methods}
}
\description{
#' Construct a single row summary "glance" of a model, fit, or other
object
}
\details{
Returns a tibble specifying the number of parameters (df),
log-likelihood (logLik), Akaike's Information Criterion (AIC) and
whether or not the model converged.
}
\section{Methods}{


\Sexpr[stage=render,results=rd]{generics:::methods_rd("glance")}

}

\examples{
expr <- "sum(dnorm(len, mu, sigma, log = TRUE))"
start <- list(mu = 20, sigma = 8)
analysis <- ml_fit(expr, start, data = datasets::ToothGrowth)
glance(analysis)
}
