% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mad2019}
\alias{mad2019}
\title{Encuesta de Condiciones de Vida (ECV 2019) Madrid}
\format{
\subsection{\code{mad2019}}{

A data frame with 1571 rows and 25 columns:
\describe{
\item{weight}{Survey weights}
\item{sex}{Sex}
\item{nadults}{Number of adults living in household when indivual was 14}
...
}
}
}
\source{
\url{https://www.ine.es/dyngs/INEbase/es/operacion.htm?c=Estadistica_C&cid=1254736176807&menu=ultiDatos&idp=1254735976608}
}
\usage{
mad2019
}
\description{
A subset of the spanish ECV 2019 survey which is part of EUSILC.
The subset restricts data to Madrid and is cleaned to have only
data on income and circumstances.
}
\keyword{datasets}
