% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{make_recs}
\alias{make_recs}
\title{Recipe constructor}
\usage{
make_recs(formula, data)
}
\arguments{
\item{formula}{a model formula. No in-line function should be used here, and
no minus signs are allowed. For high dimensional sets avoid this function.}

\item{data}{A data frame or tibble.}
}
\description{
Make preprocessor specifations using the \code{recipes} package. Each
recipe specifies different steps to be applied during the analysis process,
depending on the type of model to be used.
}
