% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_6_resample_train.R
\name{ml_6_resample_train}
\alias{ml_6_resample_train}
\title{Create samples of training data for varying model training needs}
\usage{
ml_6_resample_train(data_sets_list, group_by = NULL, sample_train = NULL,
  replace = FALSE, output = "list", add = FALSE, seed = NULL)
}
\arguments{
\item{data_sets_list}{A list containing at least one data frame \code{.train}
and possibly other data frames \code{.test1} and \code{.test2}, as well as 
other dot prefixed elements specified earlier in the pipeline. Typically the
\code{.cv_folds} element should already be calcualted.}

\item{group_by}{One or more grouping vectors from the .train element or the 
dot prefixed elements within \code{data_sets_list}, which specifies the 
lower hierarchies, to be sampled from.}

\item{sample_train}{An integer containing the number of samples from each group.}

\item{replace}{Logical (default \code{FALSE}). Should sampling be with replacement?}

\item{output}{A character vector specifying the output type. One can choose between
list, tibble, data.table and data.frame. Other formats like sparse matrices
might be implemented in the future.}

\item{add}{Per default new resamples by iterative calls will overwrite those
from the first call. If you instead want to add an additional resamplings,
 set \code{add} to \code{TRUE}.}

\item{seed}{The seed (integer) used to create the random samples.}
}
\value{
A list containing the train and test sets as well as a .resample_train
and other dot prefixed elements as well.
}
\description{
Create samples of training data for varying model training needs
}
\examples{
ml_6_resample_train(data = iris)

}
\seealso{
\href{https://github.com/Tazinho/ml}{ml on github}.
}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
