% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_7_resample_test.R
\name{ml_7_resample_test}
\alias{ml_7_resample_test}
\title{Create samples of test data for varying model validation needs}
\usage{
ml_7_resample_test(data_sets_list, group_by = NULL, sample_train = NULL,
  replace = FALSE, seed = NULL)
}
\arguments{
\item{data_sets_list}{A list containing at least one data frame \code{.train}
and possibly other data frames \code{.test1} and \code{.test2}, as well as 
other dot prefixed elements specified earlier in the pipeline. Typically the
\code{.cv_folds} element should already be calcualted.}

\item{group_by}{One or more grouping vectors from the .train element or the 
dot prefixed elements within \code{data_sets_list}, which specifies the 
lower hierarchies, to be sampled from.}

\item{sample_train}{An integer containing the number of samples from each group.}

\item{replace}{Logical (default \code{FALSE}). Should sampling be with replacement?}

\item{seed}{The seed (integer) used to create the random samples.}
}
\value{
A list containing the train and test sets as well as a .resample_test
and other dot prefixed elements as well.
}
\description{
Create samples of test data for varying model validation needs
}
\examples{
ml_7_resample_test(data = iris)

}
\seealso{
\href{https://github.com/Tazinho/ml}{ml on github}.
}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
