% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_tidymodels.R
\name{ML_add}
\alias{ML_add}
\title{construct rsample splits from data in an environment}
\usage{
ML_add(env, ...)
}
\arguments{
\item{env}{environment}

\item{...}{arguments passed to initial_split}
}
\description{
construct rsample splits from data in an environment
}
