% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{ml_analyse}
\alias{ml_analyse}
\title{Maximum Likelhood Analysis}
\usage{
ml_analyse(expr, pars, data)
}
\arguments{
\item{expr}{A string of R code that evaluates to the log-likelihood.}

\item{pars}{An named list of the parameter initial values (coerced to an nlist object).}

\item{data}{An named list of the data (coerced to an nlist object).}
}
\value{
An ml_analysis object.
}
\description{
Maximum Likelhood Analysis
}
