% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-table.R
\name{ml_coef_table}
\alias{ml_coef_table}
\title{Coefficient Table}
\usage{
ml_coef_table(x, conf_level = 0.95)
}
\arguments{
\item{x}{An ml_analysis Object}

\item{conf_level}{A number between 0 and 1 specifying the confidence level.}
}
\value{
A tibble of the coefficient table.
}
\description{
Coefficient Table
}
\examples{
 set.seed(101)
 object <- ml_analyse("sum(dpois(x, exp(log_lambda), log = TRUE))",
            pars = list(log_lambda = 0), data = list(x = rpois(10, 1.5)))
 ml_coef_table(object)
}
