% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{ml_fit}
\alias{ml_fit}
\title{Fit an R Expression based Maximum Likelihood Model to Data}
\usage{
ml_fit(expr, start, data)
}
\arguments{
\item{expr}{A string of R code that evaluates to the log-likelihood.}

\item{start}{An named list of the starting values for the parameter terms
(coerced to an nlist object). Missing values are fixed at 0.}

\item{data}{An named list or data frame of the data (coerced to an nlist object).}
}
\value{
A fitted ml_analysis object.
}
\description{
Fit an R Expression based Maximum Likelihood Model to Data
}
\examples{
expr <- "sum(dnorm(len, mu, sigma, log = TRUE))"
start <- list(mu = 20, sigma = 8)
data <- datasets::ToothGrowth
analysis <- ml_fit(expr, start = start, data = data)
analysis
}
