% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_9_model.R
\name{ml_model}
\alias{ml_model}
\title{Specify models and parameters to train, possible on a grid of hyper parameters.}
\usage{
ml_model(data_sets_list, model = NULL, parameters = NULL, formula = NULL,
  hyper_parameters = NULL, grid_search_tries = NULL, parallel = NULL,
  output = "list", add = FALSE)
}
\arguments{
\item{data_sets_list}{A list containing at least one data frame \code{.train}
and possibly other data frames \code{.test1} and \code{.test2}, as well as 
other dot prefixed elements specified earlier in the pipeline.}

\item{model}{Name of (at least) one modelling function, prefered in
package::name syntax. A list containing all possible models will be 
available soon.}

\item{parameters}{A list containing typical parameters like \code{type = response}.}

\item{formula}{Model formula. Doesn't need to be specified if \code{features}
or \code{target} where already specified within the pipeline.}

\item{hyper_parameters}{A list of hyperparameters. If more than one model is
supplied within one function call, the list needs to have one list element
per model.}

\item{grid_search_tries}{If \code{NULL} (default), the full grid is calculated.
If an integer is supplied, a random grid search is initiated, running the
regarding rounds.}

\item{parallel}{This option is not implemented yet. It should be possible
to specify the hierarchies of parallelisation.}

\item{output}{A character vector specifying the output type. One can choose between
list, tibble, data.table and data.frame. Other formats like sparse matrices
might be implemented in the future.}

\item{add}{Per default new models and parameters by iterative calls will 
overwrite those from the first call. If you instead want to add an additional
model, set \code{add} to \code{TRUE}.}
}
\value{
A list containing the train and test sets as well as a .target
element and other elements prefixed by a dot, if supplied earlier in the
pipeline.
}
\description{
Specify models and parameters to train, possible on a grid of hyper parameters.
}
\seealso{
\href{https://github.com/Tazinho/ml}{ml on github}.
}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
