% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLcv.R
\name{MLcv}
\alias{MLcv}
\title{Perform CV on list of machine learners}
\usage{
MLcv(
  X,
  Y,
  ML = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "Logit_lasso", "OLS", "grf",
    "OLSensemble"),
  OLSensemble,
  SL.library,
  Kcv = 5,
  rf.cf.ntree = 500,
  rf.depth = NULL,
  mtry = max(floor(ncol(X)/3), 1),
  ensemblefolds = 2,
  polynomial.Lasso = 1,
  polynomial.Ridge = 1,
  polynomial.Logit_lasso = 1,
  polynomial.OLS = 1,
  xgb.nrounds = 200,
  xgb.max.depth = 6,
  cb.iterations = 500,
  cb.depth = 6,
  verbose = FALSE,
  weights = NULL
)
}
\arguments{
\item{X}{is a dataframe containing all the features on which the
model was estimated}

\item{Y}{is a vector containing the labels for which the model
was estimated}

\item{ML}{string vector specifying which machine learners to use}

\item{OLSensemble}{is a string vector specifying which learners
should be used in OLS ensemble method}

\item{SL.library}{is a string vector specifying which learners
should be used in SuperLearner}

\item{Kcv}{number of folds in cross-validation}

\item{rf.cf.ntree}{how many trees should be grown when using RF or CIF}

\item{rf.depth}{how deep should trees be grown in RF (NULL is default from ranger)}

\item{mtry}{how many variables to consider at each split in RF}

\item{ensemblefolds}{is an integer specifying how many folds to use in ensemble
methods such as OLSensemble or SuperLearner}

\item{polynomial.Lasso}{degree of polynomial to be fitted when using Lasso.
1 just fits the input X. 2 squares all variables and adds
all pairwise interactions. 3 squares and cubes all variables and adds all
pairwise and threewise interactions...}

\item{polynomial.Ridge}{degree of polynomial to be fitted when using Ridge,
see polynomial.Lasso for more info.}

\item{polynomial.Logit_lasso}{degree of polynomial to be fitted when using Logit_lasso,
see polynomial.Lasso for more info.}

\item{polynomial.OLS}{degree of polynomial to be fitted when using OLS,
see polynomial.Lasso for more info.}

\item{xgb.nrounds}{is an integer specifying how many rounds to use in XGB}

\item{xgb.max.depth}{is an integer specifying how deep trees should be grown in XGB}

\item{cb.iterations}{The maximum number of trees that can be built in CB}

\item{cb.depth}{The depth of the trees in CB}

\item{verbose}{logical specifying whether to print progress}

\item{weights}{is a vector containing survey weights adding up to 1}
}
\value{
list containing ML attaining minimum RMSE and RMSE
}
\description{
\code{MLcv} computes cross-validated RMSE for a list including up to
Lasso, Ridge,
Random Forest, Conditional Inference Forest, Logit lasso,
Extreme Gradient Boosting and Catboosting. Returns ML with minimum
RMSE and the RMSE.
}
