% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLtuning.R
\name{MLtuning}
\alias{MLtuning}
\title{Perform CV on list of machine learners}
\usage{
MLtuning(
  X,
  Y,
  ML = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "Logit_lasso", "OLS", "grf",
    "OLSensemble"),
  OLSensemble,
  SL.library,
  Kcv = 5,
  rf.cf.ntree.grid = c(100, 300, 500),
  rf.depth.grid = c(2, 4, 6, 10),
  mtry.grid = c(1, 3, 5),
  ensemblefolds.grid = c(2, 5),
  polynomial.grid = c(1, 2, 3),
  xgb.nrounds.grid = c(100, 200, 500),
  xgb.max.depth.grid = c(1, 3, 6),
  verbose = FALSE
)
}
\arguments{
\item{X}{is a dataframe containing all the features on which the
model was estimated}

\item{Y}{is a vector containing the labels for which the model
was estimated}

\item{ML}{string vector specifying which machine learners to use}

\item{OLSensemble}{is a string vector specifying which learners
should be used in OLS ensemble method}

\item{SL.library}{is a string vector specifying which learners
should be used in SuperLearner}

\item{Kcv}{number of folds in cross-validation}

\item{rf.cf.ntree.grid}{how many trees should be grown when using RF or CIF}

\item{rf.depth.grid}{how deep should trees be grown in RF (NULL is always tried)}

\item{mtry.grid}{how many variables to consider at each split in RF,
defaults floor(sqrt(ncol(X))) and floor(ncol(X)/3) are always tried}

\item{ensemblefolds.grid}{is an integer specifying how many folds to use in ensemble
methods such as OLSensemble or SuperLearner}

\item{polynomial.grid}{degree of polynomial to be fitted when using Lasso, Ridge,
Logit Lasso or OLS. 1 just fits the input X. 2 squares all variables and adds
all pairwise interactions. 3 squares and cubes all variables and adds all
pairwise and threewise interactions...}

\item{xgb.nrounds.grid}{is an integer specifying how many rounds to use in XGB}

\item{xgb.max.depth.grid}{is an integer specifying how deep trees should be grown in XGB}

\item{verbose}{logical specifying whether to print progress}
}
\value{
list containing ML attaining minimum RMSE and RMSE
}
\description{
\code{MLtuning} performs hyperparameter tuning through cross-validation for
Lasso, Ridge, Random Forest, Conditional Inference Forest, Logit lasso,
Extreme Gradient Boosting and Catboosting (some might be missing still)
Returns hyperparameters with minimum
RMSE and the RMSE.
}
