% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_tidymodels.R
\name{mod}
\alias{mod}
\title{Create \code{parsnip} model specifications}
\usage{
mod(fun, engine, ..., mode = "regression")
}
\arguments{
\item{fun}{a \code{parsnip} function}

\item{engine}{a \code{parsnip} engine specification}

\item{...}{arguments passed to \code{parsnip} model specificatin constructors.
Empty arguments get converted to \code{tune::tune()}}

\item{mode}{character of length 1, one of \emph{regression} or \emph{classification}}
}
\description{
Create \code{parsnip} model specifications
}
