% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modest_function.R
\name{modest}
\alias{modest}
\title{Estimate Machine Learning model}
\usage{
modest(
  X,
  Y,
  ML = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "SL"),
  ensemble = c("SL.Lasso", "SL.Ridge", "SL.RF", "SL.CIF", "SL.XGB", "SL.CB"),
  weights = NULL
)
}
\arguments{
\item{X}{is a dataframe containing all the features}

\item{Y}{is a vector containing the label}

\item{ML}{is a string specifying which machine learner to use}

\item{ensemble}{is a string vector specifying which learners
to combine in the SuperLearner. Only needed if ML = "SL".}

\item{weights}{is a vector containing survey weights adding up to 1}
}
\value{
the object that the machine learner package returns
}
\description{
\code{modest} estimates the model for a specified machine learner,
possible options are Lasso, Ridge, Random Forest, Conditional
Inference Forest, Extreme Gradient Boosting, Catboosting or any
combination of these using the SuperLearner package
}
\details{
Note that the glmnet package which implements Lasso and Ridge
does not handle factor variables (such as the ones in mad2019)
hence if X contains factors dummy encoding needs to be used. This
can be done by simply using stats::model.matrix(X) instead of X
}
\examples{
as_tibble(mad2019)
X <- dplyr::select(mad2019,-c(Y,weight))
Y <- mad2019$Y
modest(X,Y,"RF")
modest(X,Y,"XGB")
modest(stats::model.matrix(X),Y,"Lasso")

}
