% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perceptron.R
\name{perceptron}
\alias{perceptron}
\title{Main perceptron algorithm with averaging}
\usage{
perceptron(y, x, E = 100, has_bias = FALSE, seed = Sys.time())
}
\arguments{
\item{y}{response variable}

\item{x}{predictors}

\item{E}{number of epochs}

\item{has_bias}{if the bias term has been included or not (defaults to FALSE)}
}
\value{
tibble of the input file
}
\description{
Main perceptron algorithm with averaging
}
\examples{
perceptron()
}
