% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perceptron.R
\name{perceptron_fit}
\alias{perceptron_fit}
\title{Perceptron Fit Calc}
\usage{
perceptron_fit(x, w, has_bias = FALSE)
}
\arguments{
\item{x}{predictors variable}

\item{w}{weights}
}
\value{
Numeric vector of responses
}
\description{
Perceptron Fit Calc
}
\examples{
perceptron_fit()
}
