% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras.R
\name{prepare_recipe}
\alias{prepare_recipe}
\title{Define and run the preprocesing recipe.}
\usage{
prepare_recipe(data)
}
\arguments{
\item{data}{An \code{rsplit} object from \code{\link[=split_data]{split_data()}}
with the training and test customer churn data.}
}
\value{
A \code{recipe} object with the preprocessing steps
and the preprocessed test data.
}
\description{
Define a recipe to preprocess the data
and run it on the test data.
}
