% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{print.ml_cv_result}
\alias{print.ml_cv_result}
\title{Print ml_cv_result}
\usage{
\method{print}{ml_cv_result}(x, ...)
}
\arguments{
\item{x}{An ml_cv_result object}

\item{...}{Ignored}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print ml_cv_result
}
