% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{print.ml_embedder}
\alias{print.ml_embedder}
\title{Print ml_embedder}
\usage{
\method{print}{ml_embedder}(x, ...)
}
\arguments{
\item{x}{An ml_embedder object}

\item{...}{Ignored}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print ml_embedder
}
