% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{print.ml_evidence}
\alias{print.ml_evidence}
\title{Print ml_evidence}
\usage{
\method{print}{ml_evidence}(x, ...)
}
\arguments{
\item{x}{An ml_evidence object}

\item{...}{Ignored}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print ml_evidence
}
