% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{print.ml_leaderboard}
\alias{print.ml_leaderboard}
\title{Print ml_leaderboard}
\usage{
\method{print}{ml_leaderboard}(x, ...)
}
\arguments{
\item{x}{An ml_leaderboard object}

\item{...}{Ignored}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print ml_leaderboard
}
