% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{print.ml_model}
\alias{print.ml_model}
\title{Print an ml_model}
\usage{
\method{print}{ml_model}(x, ...)
}
\arguments{
\item{x}{An ml_model object}

\item{...}{Ignored}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print an ml_model
}
