% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{print.ml_tuning_result}
\alias{print.ml_tuning_result}
\title{Print an ml_tuning_result}
\usage{
\method{print}{ml_tuning_result}(x, ...)
}
\arguments{
\item{x}{An ml_tuning_result object}

\item{...}{Ignored}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print an ml_tuning_result
}
