% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{print.ml_validate_result}
\alias{print.ml_validate_result}
\title{Print ml_validate_result}
\usage{
\method{print}{ml_validate_result}(x, ...)
}
\arguments{
\item{x}{An ml_validate_result object}

\item{...}{Ignored}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print ml_validate_result
}
