% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ingestion.R
\name{read_libsvm}
\alias{read_libsvm}
\title{Read a file in LIBSVM format}
\usage{
read_libsvm(x, prefix = "F")
}
\arguments{
\item{x}{string representing full or relative path of input file}

\item{prefix}{string or character value for the predictor columns}
}
\value{
tibble of the input file
}
\description{
Read a file in LIBSVM format
}
\examples{
read_libsvm()
}
