% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_bookdown.R
\name{render_with_deps}
\alias{render_with_deps}
\title{Bookdown render with explicit defined dependencies}
\usage{
render_with_deps(index, deps)
}
\arguments{
\item{index}{file to index and render}

\item{deps}{dependencies}
}
\description{
Bookdown render with explicit defined dependencies
}
