% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results}
\alias{results}
\title{Best performing ML model in group project}
\format{
A list containing three elements
\describe{
\item{predictions}{a tibble with model performance predictions and observed events.}
\item{workflow}{a \code{workflow} object that specifies exact hyperparameter settings of the best model}
}
}
\source{
\url{https://github.com/D-Se/ML}
}
\usage{
results
}
\description{
A quick insight into the best performing model of the bunch
}
\keyword{datasets}
