% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_in_env}
\alias{set_in_env}
\title{Evaluate a set of expression and assign to an environment}
\usage{
set_in_env(env, expr, ...)
}
\arguments{
\item{env}{an environment object}

\item{expr}{an expression}

\item{...}{used to pass objects to the expression}
}
\description{
Evaluate a set of expression and assign to an environment
}
\keyword{internal}
