% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras.R
\name{test_accuracy}
\alias{test_accuracy}
\title{Test accuracy.}
\usage{
test_accuracy(churn_data, churn_recipe, model)
}
\arguments{
\item{churn_recipe}{A \code{recipe} object from \code{\link[=prepare_recipe]{prepare_recipe()}}.}
}
\value{
Classification accuracy of a trained Keras model on the test
dataset.
}
\description{
Compute the classification accuracy of a trained Keras model
on the test dataset.
}
\examples{
library(keras)
library(recipes)
library(rsample)
library(tidyverse)
library(yardstick)
data <- split_data("data/customer_churn.csv")
recipe <- prepare_recipe(data)
model <- train_model(recipe, 16, 16, "relu", "relu", "sigmoid")
test_accuracy(data, recipe, model)
}
