% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras.R
\name{test_model}
\alias{test_model}
\title{Benchmark a Keras model.}
\usage{
test_model(
  data,
  recipe,
  units1 = 16,
  units2 = 16,
  act1 = "relu",
  act2 = "relu",
  act3 = "sigmoid"
)
}
\arguments{
\item{data}{a matrix.}

\item{recipe}{A \code{recipe} object from \code{\link[=prepare_recipe]{prepare_recipe()}}.}

\item{units1}{Number of neurons in the first layer.}

\item{units2}{Number of neurons in the second layer.}

\item{act1}{Activation function for layer 1.}

\item{act2}{Activation function for layer 2.}

\item{act3}{Activation function for layer 3.}
}
\value{
A data frame with one row and the following columns:
\itemize{
\item \code{accuracy}: classification accuracy on the test dataset.
\item \code{units1}: number of neurons in layer 1.
\item \code{units2}: number of neurons in layer 2.
\item \code{act1}: number of neurons in layer 1.
\item \code{act2}: number of neurons in layer 2.
\item \code{act3}: number of neurons in layer 3.
}
}
\description{
Define, compile, and train a Keras model on the training
dataset. Then, benchmark it on the test dataset and return summaries.
}
\details{
The first time you run Keras in an R session,
TensorFlow usually prints verbose ouput such as
"Your CPU supports instructions that this TensorFlow binary was not compiled to use:"
and "OMP: Info #171: KMP_AFFINITY:". You can safely ignore these messages.
}
