% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras.R
\name{train_best_model}
\alias{train_best_model}
\title{Retrain the best model.}
\usage{
train_best_model(best_run, churn_recipe)
}
\arguments{
\item{best_run}{blabla.}

\item{churn_recipe}{a recipe.}
}
\value{
A trained Keras model.
}
\description{
After we find the model with the best accuracy,
retrain it and return the trained model given a row of output
from \code{\link[=test_model]{test_model()}}.
}
\details{
The first time you run Keras in an R session,
TensorFlow usually prints verbose ouput such as
"Your CPU supports instructions that this TensorFlow binary was not compiled to use:"
and "OMP: Info #171: KMP_AFFINITY:". You can safely ignore these messages.
}
\examples{
library(keras)
library(recipes)
library(rsample)
library(tidyverse)
data <- split_data("data/customer_churn.csv")
recipe <- prepare_recipe(data)
run <- test_model(data, recipe, 16, 16, "relu", "relu", "sigmoid")
train_best_model(run, recipe)
}
