% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_tidymodels.R
\name{add_splits}
\alias{add_splits}
\title{construct rsample splits from data in an environment}
\usage{
add_splits(..., env = new.env())
}
\arguments{
\item{...}{arguments passed to initial_split}

\item{env}{environment}
}
\description{
construct rsample splits from data in an environment
}
