% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance.ml_analysis}
\alias{glance.ml_analysis}
\title{Glance at an ml_analysis Object}
\usage{
\method{glance}{ml_analysis}(x, ...)
}
\arguments{
\item{x}{An ml_analysis object.}

\item{...}{Unused.}
}
\value{
A tibble specifying the number of parameters (df),
log-likelihood (logLik), Akaike's Information Criterion (AIC) and
whether or not the model converged.
}
\description{
Construct a summary "glance" of an ml_analysis object.
}
\examples{
expr <- "sum(dnorm(len, mu, sigma, log = TRUE))"
start <- list(mu = 20, sigma = 8)
analysis <- ml_fit(expr, start, data = datasets::ToothGrowth)
glance(analysis)
}
