% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML-package.R
\docType{package}
\name{ML-package}
\alias{ML}
\alias{ML-package}
\title{ML: Machine Leanring targets-bookdown pipeline}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A maximum reproducible example of a machine learning pipeline, package-based workflow implemented for the Nanjing University Spring 2022 Machine Learning 01 class (020205D17).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/D-Se/ML}
  \item Report bugs at \url{https://github.com/D-Se/ML/issues}
}

}
\author{
\strong{Maintainer}: Donald Seinen \email{2869830202@qq.com} (\href{https://orcid.org/0000-0002-4343-5851}{ORCID})

}
\keyword{internal}
