% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml-package.R
\docType{package}
\name{ml-package}
\alias{ml}
\alias{ml-package}
\title{ml: Maximum Likelihood Analysis using R Expressions}
\description{
Performs Maximum Likelihood analysis using R expressions.
    It is designed to be simple but flexible.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/ml}
  \item Report bugs at \url{https://github.com/poissonconsulting/ml/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
