% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_5_target.R
\name{ml_5_target}
\alias{ml_5_target}
\title{Specify a target variable for a machine learning model}
\usage{
ml_5_target(data_sets_list, target, pos_class = NULL, output = "list",
  add = FALSE)
}
\arguments{
\item{data_sets_list}{A list containing at least one data frame \code{.train}
and possibly other data frames \code{.test1} and \code{.test2}, as well as 
other dot prefixed elements specified earlier in the pipeline.}

\item{target}{A character vector or NSE containing the target name, which 
must be a column name of training and test sets within \code{data_sets_list}. 
Alternatively a formula can be supplied, the right hand side elements are
automaitcally known as .target and a call to \code{ml_target} is not 
necessary anymore. Typically only one target should be set, but it is possible
to create many models at once. Currently just one column name supplied as 
character is implemented.}

\item{pos_class}{short for positive class, relevant for binary classification
problems. This should be supplied as a character, but may also be supplied as
as integer, or logical, if the target is encoded in this way. If \code{pos_class}
is not supplied, the positive class will be assumed to be the less frequent class
on the whole training set within the \code{ml_model} step.}

\item{output}{A character vector specifying the output type. One can choose between
list, tibble, data.table and data.frame. Other formats like sparse matrices
might be implemented in the future.}

\item{add}{Per default a new target by iterative calls will overwrite that from
the first call. If you instead want to add an additional target, set \code{add}
to \code{TRUE}.}
}
\value{
A list containing the train and test sets as well as a .target
element and other elements prefixed by a dot, if supplied earlier in the
pipeline.
}
\description{
Specify a target variable for a machine learning model
}
\examples{
ml_5_target(data_sets_list = list(.train = iris), target = c("Petal.Width"))

}
\seealso{
\href{https://github.com/Tazinho/ml}{ml on github}.
}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
