% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{ml_analyse}
\alias{ml_analyse}
\title{Analyse R Function}
\usage{
ml_analyse(expr, pars, data)
}
\arguments{
\item{expr}{A string of the R code that uses pars and data to calculate the negative log-likelihood (nll).}

\item{pars}{An nlist object of the parameters to estimate.}

\item{data}{An nlist object of the data.}
}
\description{
Analyse R Function
}
