% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-table.R
\name{ml_coef_table}
\alias{ml_coef_table}
\title{Coefficient Table}
\usage{
ml_coef_table(x, conf_level = 0.95)
}
\arguments{
\item{x}{An ml_analysis Object}

\item{conf_level}{A number between 0 and 1 specifying the confidence level.}
}
\value{
A tibble of the coefficient table.
}
\description{
Coefficient Table
}
\examples{
template <- "sum(dnorm(len, mu, sigma, log = TRUE))"
pars <- list(mu = 20, sigma = 8)
data <- datasets::ToothGrowth
analysis <- ml_analyse(template, pars = pars, data = data)
ml_coef_table(analysis)
}
