% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converged.R
\name{ml_converged}
\alias{ml_converged}
\title{Converged}
\usage{
ml_converged(x)
}
\arguments{
\item{x}{An ml_analysis object.}
}
\value{
A flag indicating whether the analysis converged.
}
\description{
Test if an ml_analysis object converged.
}
\examples{
analysis <- ml_analyse("sum(dpois(x, lambda, log = TRUE))",
  pars = list(lambda = 1), data = list(x = rpois(1000, 1.5)))
ml_converged(analysis)
}
