% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converged.R
\name{ml_converged}
\alias{ml_converged}
\title{Converged}
\usage{
ml_converged(x)
}
\arguments{
\item{x}{An ml_analysis object.}
}
\value{
A flag indicating whether the analysis converged.
}
\description{
Test if an ml_analysis object converged.
}
\examples{
template <- "sum(dnorm(len, mu, sigma, log = TRUE))"
pars <- list(mu = 20, sigma = 8)
data <- datasets::ToothGrowth
analysis <- ml_analyse(template, pars = pars, data = data)
ml_converged(analysis)
}
