% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_split.R
\name{ml_split}
\alias{ml_split}
\title{Splitting datasets into training and test sets for machine learning}
\usage{
ml_split(data, train = NULL, test1 = NULL, test2 = NULL,
  stratify = NULL, seed = NULL)
}
\arguments{
\item{data}{A data frame}

\item{train}{This argument specifies the part of the \code{data}, which should
be used to train a model. It can be provided as a percentage between 0 and 1,
or an index. The index can be supplied as a logical, an integer or as an
expression, which becomes evaluated in the data (for example a column of
\code{data}). If \code{data} is NULL, a data frame must be supplied instead.}

\item{test1}{This argument specifies the part of the \code{data}, which should
be used to test the model, which was trained on \code{train}.
It can be supplied in the same way as \code{train}}

\item{test2}{This argument specifies the part of the \code{data}, which should
be used to test a model, which was trained on \code{train} a second.time
It can be supplied in the same way as \code{train} and \code{test1}}

\item{stratify}{The name of one or more columns in \code{data}. When
\code{train}, \code{test1} and \code{test2} are supplied, as percentages, 
the resulting proportions of the supplied columns in the created training and
test sets will be the same.}

\item{seed}{The seed (integer) used to create the random samples, if
\code{train}, \code{test1} and \code{test2} where supplied as percentages.}
}
\value{
A list containing the train and test sets.
}
\description{
Splitting datasets into training and test sets for machine learning
}
\note{
When \code{train} is supplied as a percentage, the rest of the \code{data}
will be used as \code{train1}. If both are supplied as percentages, the rest
of the data will be used as \code{train2}.
}
\examples{
ml_split(data = iris, train = c(T, F, F), test1 = c(F, T, F), test2 = c(F, F, T))

}
\seealso{
\href{https://github.com/Tazinho/ml}{ml on github}.
}
\author{
Malte Grosser, \email{malte.grosser@gmail.com}
}
\keyword{utilities}
