% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modest.R
\name{modest}
\alias{modest}
\title{Estimate Machine Learning model}
\usage{
modest(
  X,
  Y,
  ML = c("Lasso", "Ridge", "RF", "CIF", "XGB", "CB", "SL"),
  ensemble = c("SL.Lasso", "SL.Ridge", "SL.RF", "SL.CIF", "SL.XGB", "SL.CB"),
  weights = NULL
)
}
\arguments{
\item{X}{is a dataframe containing all the features}

\item{Y}{is a vector containing the label}

\item{ML}{is a string specifying which machine learner to use}

\item{ensemble}{is a string vector specifying which learners
to combine in the SuperLearner. Only needed if ML = "SL".}

\item{weights}{is a vector containing survey weights adding up to 1}
}
\value{
the object that the machine learner package returns
}
\description{
\code{modest} estimates the model for a specified machine learner,
possible options are Lasso, Ridge, Random Forest, Conditional
Inference Forest, Extreme Gradient Boosting, Catboosting or any
combination of these using the SuperLearner package
}
\details{
Note that the glmnet package
which implements Lasso and Ridge does not handle factor variables
(such as the ones in mad2019), hence for this machine learners,
modest turns X into model.matrix(~.,X) which will perform dummy
encoding on factor variables.
}
\examples{
X <- dplyr::select(mad2019,-Y)
Y <- mad2019$Y
modest(X,Y,"RF")
modest(X,Y,"XGB")
modest(X,Y,"Lasso")
modest(X,Y,"SL",
ensemble = c("SL.Lasso","SL.Ridge","SL.RF","SL.CIF","SL.XGB","SL.CB"))

}
