% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_keras.R
\name{test_accuracy}
\alias{test_accuracy}
\title{Test accuracy.}
\usage{
test_accuracy(data, recipe, model)
}
\arguments{
\item{data}{a matrix.}

\item{recipe}{a \code{recipe} object from the \code{recipes} packages.}

\item{model}{a \code{keras} model to assess.}
}
\value{
Classification accuracy of a trained Keras model on the test
dataset.
}
\description{
Compute the classification accuracy of a trained Keras model
on the test dataset.
}
