% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-table.R
\name{tidy.ml_analysis}
\alias{tidy.ml_analysis}
\title{Turn an ml_analysis Object into a tidy tibble}
\usage{
\method{tidy}{ml_analysis}(x, constant = FALSE, conf_level = 0.95, ...)
}
\arguments{
\item{x}{An ml_analysis Object}

\item{constant}{A logical scalar specifying whether to include constant terms.}

\item{conf_level}{A number between 0 and 1 specifying the confidence level.}

\item{...}{Unused}
}
\value{
A data.frame of the coefficient tables with their
estimates, standard errors (sd), lower and upper confidence limits
and surprisal values (svalue).
}
\description{
Turn an ml_analysis Object into a tidy tibble
}
\examples{
expr <- "sum(dnorm(len, mu, sigma, log = TRUE))"
pars <- list(mu = 20, sigma = 8)
data <- datasets::ToothGrowth
analysis <- ml_analyse(expr, pars = pars, data = data)
tidy(analysis)
}
