% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_keras.R
\name{train_best_model}
\alias{train_best_model}
\title{Retrain the best model.}
\usage{
train_best_model(best_run, recipe)
}
\arguments{
\item{best_run}{blabla.}

\item{recipe}{a recipe.}
}
\value{
A trained Keras model.
}
\description{
After we find the model with the best accuracy,
retrain it and return the trained model given a row of output
from \code{\link[=test_model]{test_model()}}.
}
\details{
The first time you run Keras in an R session,
TensorFlow usually prints verbose ouput such as
"Your CPU supports instructions that this TensorFlow binary was not compiled to use:"
and "OMP: Info #171: KMP_AFFINITY:". You can safely ignore these messages.
}
